Complete exercise 04-chain-selectors of TOP
Part of the assignment for The Odin Project Foundations Course CSS
Foundations Intro to CSS.

A part of the challenge here was determining whether or not to include
the HTML height and width attributes. I concluded it should be included
as recommended by the lesson. Concisely, without those attributes, there
may be a layout shift.

Layout shift occurs when there is a repainted of the page where the
contents are changed in position or size, which in turn, affects the
position of other content.

In the case of an absent of image when the page is initially painted,
the contents must be shifted to the right or downwards to accommodate
the image when it is finally loaded.

Even if the image dimensions are to be changed using CSS, the layout
shift will not be as drastic. Also, any relative changes to the
dimensions of the image may be calculated before the image even loads
since the HTML width and height attributes provide an aspect ratio.
Without those attributes, any calculation would have to wait for the
image to load to obtain its intrinsic aspect ratio.

Source:
https://developer.mozilla.org/en-US/docs/Learn/Performance/Multimedia