Complete exercise 06-cascade-fix of TOP
Part of the assignment for The Odin Project Foundations Course CSS
Foundations Intro to CSS.

The cascade does use the number of most specific selectors to determine
which rule to apply when they target the same element, which is known as
a conflicting rule. That is, the conflicting rule with the most specific
selectors in a chaining selector or a selector with a combinator will be
applied.

The selector type does not matter when determining specificity. Given
a descendant selector having an id selector and class selector, in
either order and given a chaining selector composed of an id selector
and a type selector, the descendant selector's rule would be applied
to any elements target by the two selectors. On the other hand, if the
descendant were to have the id selector and type selector and the
chaining selector were to have the id selector and the class selector,
the chaining selector would win. This is because they each have only one
id selector. But the one with the class selector is applied since the
class selector is considered to have more specificity than the selector
with less than that number of class selectors, which is 0 in this case.
if the one that had 0 class selectors had several type selectors it
still would not be applied.

However, grouping selectors have the same specificity as if the rule
were repeated as many times as the number of selectors that were
grouped, i.e. separated by commas, but each rule use a distinct selector
from the grouped selectors.