Complete exercise 02-flex-header of TOP
The one difference between my solution and the solution found in the
upstream repo is I used margin property on the items to add space
between the items and the header instead of using the padding property
on the header.

Even though my solution achieved the same effect in this particular
example, if more items were to be added to the edges, the margin
property would have to be transferred to them from the respective side.
Or if another item ended up with more height than the logo for whatever
reason, the margin would have to be added to that item.

That is not the case if padding were added to the header. It would make
it simpler to change the items in the header.