Complete exercise 03-flex-header-2 of TOP
Missed the margin on the ul element so the height for the header was
more than the solution's. It might have been picked up if I inspected
the web dev tools.

Even though it does not fulfill the requirements and I consider
extraneous to the exercise, the profile pic element was made to prevent
shrinking by using flex-shrink: 0; rule declaration. Alternatively, use
min-width instead of or in addition to the width property.

Another difference is I combined some of the rules for the ul and
left and right flex items but with an additional property of
align-items: center; targeting ul that that serves no purpose here.
Only if the ul had its height changed where there is positive available
space in the vertical would that.