Complete exercise 04-flex-information of TOP
The text-align property is inherited by all descendants. So by
targeting the body of the web page, all visisble elements will inherit
the same text-align property unless another rule targets the descendants
explicitly as indicated by the selector of the rule.

max-width and width seem interchangeable in this exercise and thats how
my solution differs. The items will not grow to fill any positive
available space as they have the default initial value of flex-grow: 0;

Im unsure how the text would not wrap. That is how would it overflow
unless the items had a fixed height, forcing the text to overflow out
of the flex item.