Complete exercise 06-flex-layout of TOP
The biggest difference between my solution and the given solution is
the given solution allows the content to fill up any positive available
space and applies justify-content: center; to center the contents
vertically.

On the other hand, my solution uses justify-content: space-between;
on the body so the space between the header and content and the content
and the footer are equal in quantity, forcing the minimally sized
content to center vertically along the page. Then there is no reason to
apply any justify-content as it would be redundant. There is no positive
available space to place between the flex items.