Complete exercise 07-flex-layout-2 of TOP
As with my other solutions, there are differences between my solution
and the given solution.

*I include a min-width for the header so the text does not wrap and
instead the user must scroll to read the entire header

*I add the box-sizing: border-box; property-value to target every
element

*I assign the flex property to the none keyword which is equivalent to
flex: 0 0 auto; instead of just leaving the defaults and simply
assigning the flex-shrink property to 0; it targets the sidebar

*I set the flex-grow to 1 and flex-shrink to 0 in the style targeting
the body which I have instead have a class name of content; then the
flex-basis is auto which in this case takes on the minimum height of
the body and so the body does not shrink below that minimum; the
minimum height ends up being the height of the padding of the body and
for any given heights from the flex items themselves including stacked
flex-items as is the case for flex-wrap: wrap; is in effect and there
are multiple flex lines

*To create spacing between the links in the sidebar, the ul element
uses the flex-box layout just to utilize the gap property; the solution
simply uses the margin-bottom property on the default flow layout.